pyproject.toml
[project]
name = "023-rakathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "ipython>=9.1.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "uvicorn>=0.34.0",
    "langgraph>=0.2.6",
    "python-dotenv>=1.0.1",
    "langgraph-cli[inmem]>=0.2.3",
    "langchain-openai>=0.3.12",
    "langchain-anthropic>=0.3.10",
    "langsmith[pytest]>=0.3.30",
    "pytest-asyncio>=0.26.0",
    "anyio>=4.9.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
